# docs/swagger.yaml
definitions:
  models.Employee:
    properties:
      email:
        type: string
      first_name:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
info:
  contact: {}
paths:
  /employees/:
    get:
      description: Get a list of all employees
      operationId: get-employees
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with a list of Employees
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
      summary: Get a list of employees
    post:
      consumes:
      - application/json
      description: Create a new employee
      operationId: create-employee
      parameters:
      - description: Employee object
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      responses:
        "201":
          description: Successful response with a created Employee
          schema:
            $ref: '#/definitions/models.Employee'
      summary: Create a new employee
  /employees/{id}:
    delete:
      description: Delete an employee by ID
      operationId: delete-employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content" "Successful response with no content
      summary: Delete an employee by ID
    get:
      description: Get an employee by ID
      operationId: get-employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with an Employee
          schema:
            $ref: '#/definitions/models.Employee'
      summary: Get an employee by ID
    put:
      consumes:
      - application/json
      description: Update an employee by ID
      operationId: update-employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee object
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      responses:
        "200":
          description: Successful response with an updated Employee
          schema:
            $ref: '#/definitions/models.Employee'
      summary: Update an employee by ID
swagger: "2.0"